+++ include/myhtmlpp/attribute.hpp
@@ -31,6 +31,14 @@ public:
     Attribute& operator=(Attribute&& other) noexcept;
 
     /**
+     * @brief Returns whether the raw myhtml_tree_attr pointer is not nullptr.
+     *
+     * @return true if `good()` returns true, false otherwise.
+     * @see Attribute::good
+     */
+    explicit operator bool() const noexcept;
+
+    /**
      * @brief Check if two attributes are equal.
      *
      * @return Whether both attributes point to the same myhtml struct;