+++ include/myhtmlpp/tree.hpp
@@ -18,13 +18,11 @@ public:
     /**
      * @brief Tree constructor.
      *
-     * Initialises m_raw_myhtml with `raw_myhtml` and
-     * m_raw_tree with `raw_tree`.
+     * Initialises m_raw_tree with `raw_tree`.
      *
-     * @param raw_myhtml A pointer to a myhtml struct.
      * @param raw_tree A pointer to a myhtml_tree struct.
      */
-    Tree(myhtml_t* raw_myhtml, myhtml_tree_t* raw_tree);
+    explicit Tree(myhtml_tree_t* raw_tree);
 
     /**
      * @brief Tree destructor.
@@ -40,9 +38,17 @@ public:
     Tree& operator=(Tree&& other) noexcept;
 
     /**
-     * @brief Check if myhtml and myhtml_tree pointers are not nullptr.
+     * @brief Returns whether the raw myhtml_tree pointer is not nullptr.
      *
-     * @return Whether m_raw_myhtml and m_raw_tree are != nullptr.
+     * @return true if `good()` returns true, false otherwise.
+     * @see Tree::good
+     */
+    explicit operator bool() const noexcept;
+
+    /**
+     * @brief Check if myhtml_tree pointer is not nullptr.
+     *
+     * @return Whether m_raw_tree is != nullptr.
      */
     [[nodiscard]] bool good() const;
 
@@ -75,8 +81,11 @@ public:
     [[nodiscard]] Node body_node() const;
 
     /**
-      * @brief Returns all nodes in the subtree of `scope_node`
-      */
+     * @brief Returns all nodes in the subtree of `scope_node`
+     *
+     * @param scope_node The scope node of the subtree.
+     * @return A vector of all nodes in the subtree of `scope_node`.
+     */
     [[nodiscard]] std::vector<Node> scope(const Node& scope_node) const;
 
     /**
@@ -88,7 +97,7 @@ public:
 
     /**
      * @brief Returns all nodes in the tree that match the css selector
-     * `selector`.
+     *        `selector`.
      *
      * @param selector The css selector.
      * @return A vector of all nodes in the tree that match `selector`.
@@ -104,17 +113,36 @@ public:
      */
     [[nodiscard]] std::vector<Node> find_by_tag(const std::string& tag) const;
 
+    /**
+     * @brief Returns all nodes in the subtree of `scope_node` where the tag
+     *        matches `tag`.
+     *
+     * @param tag The tag to search.
+     * @param scope_node The scope node where searching starts.
+     * @return A vector of all nodes in the subtree of `scope_node` where
+     *         `tag_name()` returns `tag`.
+     */
     [[nodiscard]] std::vector<Node> find_by_tag(const std::string& tag,
                                                 const Node& scope_node) const;
     /**
      * @brief Returns all nodes in the tree where the tag matches `tag`.
      *
      * @param tag The tag to search.
+     * @param scope_node The scope node where searching starts.
      * @return A vector of all nodes in the tree where
      *         `tag_id()` returns `tag`.
      */
     [[nodiscard]] std::vector<Node> find_by_tag(TAG tag) const;
 
+    /**
+     * @brief Returns all nodes in the subtree of `scope_node` where the tag
+     *        matches `tag`.
+     *
+     * @param tag The tag to search.
+     * @param scope_node The scope node where searching starts.
+     * @return A vector of all nodes in the subtree of `scope_node` where
+     *         `tag_id()` returns `tag`.
+     */
     [[nodiscard]] std::vector<Node> find_by_tag(TAG tag,
                                                 const Node& scope_node) const;
 
@@ -127,6 +155,15 @@ public:
      */
     [[nodiscard]] std::vector<Node> find_by_class(const std::string& cl) const;
 
+    /**
+     * @brief Returns all nodes in the subtree of `scope_node` where the class
+     *        matches `cl`.
+     *
+     * @param cl The value of the class to search.
+     * @param scope_node The scope node where searching starts.
+     * @return A vector of all nodes in the subtree of `scope_node` that have an
+     *         attribute where key() returns \"class\" and value() returns `cl`.
+     */
     [[nodiscard]] std::vector<Node> find_by_class(const std::string& cl,
                                                   const Node& scope_node) const;
 
@@ -139,6 +176,15 @@ public:
      */
     [[nodiscard]] std::vector<Node> find_by_id(const std::string& id) const;
 
+    /**
+     * @brief Returns all nodes in the subtree of `scope_node` where the id
+     *        matches `id`.
+     *
+     * @param id The value of the id to search.
+     * @param scope_node The scope node where searching starts.
+     * @return A vector of all nodes in the subtree of `scope_node` that have an
+     *         attribute where key() returns \"id\" and value() returns `id`.
+     */
     [[nodiscard]] std::vector<Node> find_by_id(const std::string& id,
                                                const Node& scope_node) const;
 
@@ -154,6 +200,16 @@ public:
     [[nodiscard]] std::vector<Node> find_by_attr(const std::string& key,
                                                  const std::string& val) const;
 
+    /**
+     * @brief Returns all nodes in the subtree of `scope_node` that have an
+     *        attribute with key `key` and value `value`.
+     *
+     * @param key The key of the attribute.
+     * @param value The value of the attribute.
+     * @param scope_node The scope node where searching starts.
+     * @return A vector of all nodes in the subtree of `scope_node` that have an
+     *         attribute where key() returns `key` and value() returns `value`.
+     */
     [[nodiscard]] std::vector<Node> find_by_attr(const std::string& key,
                                                  const std::string& val,
                                                  const Node& scope_node) const;
@@ -259,15 +315,14 @@ public:
     [[nodiscard]] ConstIterator cend() const noexcept;
 
 private:
-    /// Pointer to the underlying myhtml struct.
-    myhtml_t* m_raw_myhtml;
-
     /// Pointer to the underlying myhtml tree struct.
     myhtml_tree_t* m_raw_tree;
 };
 
 /**
- * @brief Print the tree.
+ * @brief Returns the serialized tree.
+ *
+ * @see Tree::html
  */
 std::ostream& operator<<(std::ostream& os, const Tree& t);